requirements.txt
pdfplumber
pandas
langchain
langchain_groq
openpyxl
python-dotenv
xlsxwriter
